+++ bigbang/src/lib.rs
@@ -15,6 +15,7 @@ use dimension::Dimension;
 use node::Node;
 /*  public-facing entry points */
 pub use as_entity::AsEntity;
+pub use bigbang_derive::AsEntity;
 pub use entity::Entity;
 pub use grav_tree::GravTree;
 pub use responsive::Responsive;
